[FIX] add load_model to utils.models and use this in predict.py and app.py
