+++ src/utils/models.py
@@ -2,6 +2,26 @@ import segmentation_models_pytorch as smp
 import torch
 import torchvision.models.segmentation as models
 
+
+def load_model(checkpoint_path, num_classes=8):
+    model = _get_deeplabv3_model(num_classes)
+
+    checkpoint = torch.load(
+        checkpoint_path,
+        map_location=torch.device("cuda" if torch.cuda.is_available() else "cpu"),
+    )
+
+    state_dict = {
+        k.replace("model.", ""): v for k, v in checkpoint["state_dict"].items()
+    }
+
+    state_dict = {k: v for k, v in state_dict.items() if "aux_classifier" not in k}
+
+    model.load_state_dict(state_dict, strict=False)
+    model.eval()  # Set the model to evaluation mode
+    return model
+
+
 def get_segmentation_model(model_name, num_classes):
     if model_name == "deeplabv3plus":
         model = _get_deeplabv3plus_model(num_classes)
@@ -11,6 +31,7 @@ def get_segmentation_model(model_name, num_classes):
         raise ValueError(f"Unknown model name: {model_name}")
 
     return model
+
 
 def _get_deeplabv3plus_model(num_classes):
     model = smp.DeepLabV3Plus(
